requirements.txt
schedule~=1.2.1
PyYAML~=6.0.1
loguru~=0.7.2
fastapi~=0.110.0
openai==1.9.0
aiohttp~=3.9.3
captcha~=0.5.0
beautifulsoup4~=4.12.3
prettytable~=3.9.0
requests~=2.31.0
websocket-client~=1.7.0
websockets~=12.0
uvicorn~=0.28.0